refactor(agent-chat): move imports to top and remove redundant re-exports in a2a-ui-converter
